App/requirements.txt
autopep8==2.0.2
beautifulsoup4==4.11.2
Brotli==1.0.9
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
dash==2.7.0
dash-bootstrap-components==1.2.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-mantine-components==0.9.0
dash-table==5.0.0
et-xmlfile==1.0.1
fake-useragent==1.1.1
Flask==2.2.2
Flask-Compress==1.13
fonttools==4.39.3
gensim==4.3.1
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.7.1
mysql-connector-python==8.0.32
nltk==3.8.1
numpy==1.23.4
openpyxl==3.0.10
packaging==23.0
pandas==1.5.1
Pillow==9.4.0
pip==22.3.1
plotly==5.11.0
protobuf==3.20.3
pycodestyle==2.10.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
regex==2023.3.23
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
setuptools==65.5.1
six==1.16.0
smart-open==6.3.0
soupsieve==2.4
tenacity==8.1.0
threadpoolctl==3.1.0
tomli==2.0.1
tqdm==4.65.0
urllib3==1.26.14
Werkzeug==2.2.2
wheel==0.38.4
wordcloud==1.8.2.2
zipp==3.10.0
